+++ typescript/ccip-server/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperlane-xyz/ccip-server
 
+## 5.5.0
+
 ## 5.4.0
 
 ## 5.3.0